changes to a_gettrace.sh mki.sh search
       modified:   a_gettrace.sh
       modified:   mki.sh
       modified:   search

+++ mki.sh
@@ -3,13 +3,13 @@ OUT=out/target/product/mfld_pr2
 
 # R3 INFO #
 R3_DIR=~/r3
-R3_RELEASE=2012WW04
+R3_RELEASE=2012_WW08
 R3_URL=jfumgbuild-depot.jf.intel.com/build/eng-builds/mfld-r3/android/ice-cream-sandwich-platform/releases
 R3_MANIFEST=$R3_URL/$R3_RELEASE/manifest-$R3_RELEASE-generated.xml
 
 # R2 INFO #
 R2_DIR=~/r2
-R2_RELEASE=2012_WW04
+R2_RELEASE=2012_WW08
 R2_URL=jfumgbuild-depot.jf.intel.com/build/eng-builds/mfld-r2/android/gingerbread-platform/releases
 R2_MANIFEST=$R2_URL/$R2_RELEASE/manifest-$R2_RELEASE-generated.xml
 
@@ -20,11 +20,23 @@ print(){
 	echo "#######################################################"
 }
 
-full_andrid_build(){
+set_dir()
+{
+	if [ "$1" = "r3" ]; then
+		DIR=$R3_DIR
+	elif [  "$1" = "r2" ]; then
+		DIR=$R2_DIR
+	else
+		DIR=$R3_DIR
+	fi
 	cd $DIR
+
+}
+
+full_andrid_build(){
+	set_dir $1
         source build/envsetup.sh
         lunch mfld_pr2-eng
-        make -j4 clean
         make -j8 mfld_pr2
 }
 
@@ -49,9 +61,13 @@ build_r3(){
 	exit
 	fi
 
-	cp $DIR/$OUT/kernel_build/arch/x86/boot/bzImage  $DIR/$OUT/boot/kernel
-	$DIR/vendor/intel/support/build_boot.sh  mfld_pr2 boot.bin
-	mv  $DIR/boot.bin $DIR/$OUT/
+	
+	vendor/intel/support/mkbootimg --cmdline "init=/init pci=noearly console=ttyMFD3 console=logk0 earlyprintk=nologger loglevel=4 hsu_dma=7 kmemleak=off androidboot.bootmedia=sdcard androidboot.hardware=mfld_pr2 ip=50.0.0.2:50.0.0.1::255.255.255.0::usb0:on apic=debug"  --ramdisk $OUT/boot/ramdisk.img \
+	--kernel $OUT/kernel_build/arch/i386/boot/bzImage \
+	--output $OUT/boot.bin \
+	--product mfld_pr2 \
+	--type mos
+	
 }
 
 build_r2(){
@@ -64,7 +80,7 @@ build_r2(){
 	fi
 
 	cp $DIR/$OUT/kernel_build/arch/x86/boot/bzImage  $DIR/$OUT/boot/kernel
-	vendor/intel/support/mkbootimg --cmdline cmdline \
+	vendor/intel/support/mkbootimg --cmdline "init=/init pci=noearly console=ttyMFD3 earlyprintk=nologger loglevel=8 hsu_dma=7 kmemleak=off androidboot.bootmedia=sdcard androidboot.hardware=mfld_pr2 ip=50.0.0.2:50.0.0.1::255.255.255.0::usb0:on" \
 	--ramdisk $OUT/boot/ramdisk.img \
 	--kernel $OUT/kernel_build/arch/i386/boot/bzImage \
 	--output $OUT/boot.bin \
@@ -73,28 +89,24 @@ build_r2(){
 }
 
 reboot(){
+	print "REBOOT"
 	adb shell "update_osip --backup --invalidate 1; reboot"
+#	adb reboot-bootloader
+	
+
 }
 
-flash_boot(){
+flash_boot()
+{
+	print "FLASH"
 	sudo fastboot flash boot $DIR/out/target/product/mfld_pr2/boot.bin
 	sudo fastboot continue
 }
 
-set_dir(){
 
-	if [ "$1" = "r3" ]; then
-		DIR=$R3_DIR
-	elif [  "$1" = "r2" ]; then
-		DIR=$R2_DIR
-	else
-		DIR=$R3_DIR
-	fi
-	cd $DIR
 
-}
-
-build_kernel(){
+build_kernel()
+{
 	reboot
 	print "BUILDING $1"
 
@@ -120,13 +132,14 @@ build_kernel(){
 
 flash_kernel()
 {
+	reboot
 	set_dir $1
 	cd $DIR
-	reboot
 	flash_boot
 }
 
-flash(){
+flash()
+{
 	cd ~/DB/CURRENT/IMAGE
 	fastboot oem system /sbin/PartitionDisk.sh /dev/mmcblk0
 	
@@ -143,46 +156,37 @@ flash(){
 #GET DAILY BUILD AND FLASHIT
 get(){
 
-	cd ~/DB
-	rm -rf CURRENT
-	mkdir CURRENT
-	cd CURRENT
 	export http_proxy=
 
-	if [ -z $1 ]; then
-		echo "add ww folder example: 2012_WW04"
-		exit 1;
-	fi
-	mkdir IMAGE
-	mkdir BLANK
-
-	#GET IMAGE
 	case $1 in 
 	r2)
-		wget -r -l1 --no-parent -A"*fastboot*.zip" http://$R2_URL/$R2_RELEASE/MFLD_PRx/flash_files/build-eng/	
-		mv $R2_URL/$R2_RELEASE/MFLD_PRx/flash_files/build-eng/* .
-		rm -rf jfumgbuild-depot.jf.intel.com
-		cd IMAGE unzip ../*fastboot*.zip
-	
-		#BLANK PHONE
-		wget -r -l1 --no-parent -A"*PR3.1*.zip" http://$R2_URL/releases/$R2_RELEASE/MFLD_PRx/flash_files/blankphone/
-		mv $R2_URL/$R2_RELEASE/MFLD_PRx/flash_files/blankphone/* . 
-		rm -rf jfumgbuild-depot.jf.intel.com ;;
-	r3)
-		wget -r -l1 --no-parent -A"*PR3.1*.zip"  http://$R3_URL/$R3_RELEASE/MFLD_PRx/flash_files/build-eng/
-	        mv $R3_URL/$R3_RELEASE/MFLD_PRx/flash_files/build-eng/* .
-        	rm -rf jfumgbuild-depot.jf.intel.com
-		cd IMAGE
-		unzip ../*PR3.1*.zip
-	        #BLANK PHONE
-        	wget -r -l1 --no-parent -A"*PR3.1*.zip" http://$R3_URL/$R3_RELEASE/MFLD_PRx/flash_files/blankphone/
-	        mv $R3_URL/$R3_RELEASE/MFLD_PRx/flash_files/blankphone/* .
-        	rm -rf jfumgbuild-depot.jf.intel.com 
-	;;
-	esac
 
-	cd ../BLANK
-	unzip ../*blank*.zip
+		THIS_RELEASE=$R2_RELEASE
+		THIS_URL=$R2_URL
+		THIS_DIR=~/DB/R2/$R2_RELEASE
+		;;
+	r3)
+		THIS_RELEASE=$R3_RELEASE
+		THIS_URL=$R3_URL
+		THIS_DIR=~/DB/R3/$R3_RELEASE
+		;;
+	esac
+	
+	if [ -e $THIS_DIR ]; then
+		echo "$THIS_DIR exists."
+		exit 1;
+	fi
+	mkdir -p $THIS_DIR
+	cd $THIS_DIR
+	print "get system file."
+	wget -r -l1 --no-parent -A"*fastboot*.zip"  http://$THIS_URL/$THIS_RELEASE/MFLD_PRx/flash_files/build-eng/
+	mv $THIS_URL/$THIS_RELEASE/MFLD_PRx/flash_files/build-eng/* .
+	rm -rf jfumgbuild-depot.jf.intel.com
+	print "get PR3.1 blankphone."
+	wget -r -l1 --no-parent -A"*PR3.1*.zip" http://$THIS_URL/$THIS_RELEASE/MFLD_PRx/flash_files/blankphone/
+	mv $THIS_URL/$THIS_RELEASE/MFLD_PRx/flash_files/blankphone/* .
+	rm -rf jfumgbuild-depot.jf.intel.com
+
 }
 
 #SYNC REPO TO A RELEASE
@@ -212,39 +216,68 @@ sync_clean_build_full()
 	clean_build_full r3
 }
 
-scratch(){
-	 print "*BULDING R2 "
+scratch_r3()
+{
+	
+	print "SCRATCH R3"
 
-	rm -rf ~/r2_scratch
-	mkdir ~/r2_scratch
-	cd ~/r2_scratch
+	if [ -e ~/RELEASES/R3/$R3_RELEASE ]; then
+		print "$R3_RELEASE BUILD EXISTS PLEASE REMOVE IT."
+		exit
+	fi
 
+        mkdir -p ~/RELEASES/R3/$R3_RELEASE
+        cd ~/RELEASES/R3/$R3_RELEASE
+
+	print "REPO INIT"
+	repo init -u git://android.intel.com/manifest -b platform/android/main -m android-main
+	
+	print "GET MANIFEST"
+        wget $R3_MANIFEST
+        MANIFEST=`ls *.xml`
+        sed -i 's/jfumg-gcrmirror.jf.intel.com/ncsgit001.nc.intel.com/g' ./$MANIFEST
+
+	
+        cp  $MANIFEST ./.repo/manifests
+	print "REPO INIT MANIFEST:"
+        repo init -m  $MANIFEST
+
+	print "REPO SYNC"
+        repo sync
+
+	print "BUILD"
+        vendor/intel/support/build_all.sh -c mfld_pr2
+}
+
+scratch_r2()
+{
+	print "SCRATCH R2"
+	if [ -e ~/RELEASES/R2/$R2_RELEASE ]; then
+		print " ~/RELEASES/R2/$R2_RELEASE BUILD EXISTS PLEASE REMOVE IT."
+		exit
+	fi
+
+        mkdir -p ~/RELEASES/R2/$R2_RELEASE
+        cd ~/RELEASES/R2/$R2_RELEASE
 	repo init -u git://android.intel.com/manifest -b gingerbread -m stable
+
+	print "GET MANIFEST"
 	wget $R2_MANIFEST
-	MANIFEST="ls *.xml"
+	MANIFEST=`ls *.xml`
 	sed -i 's/jfumg-gcrmirror.jf.intel.com/ncsgit001.nc.intel.com/g' ./$MANIFEST
 	cp  $MANIFEST ./.repo/manifests
+
+	print " REPO INIT"
 	repo init -m  $MANIFEST
+	print "REPO SYNC"
 	repo sync
+	print "BUILD"
 	vendor/intel/support/build_all.sh -c mfld_pr2
 
-	print  "*BULDING R3"
-
-	rm -rf ~/r3_scratch
-        mkdir ~/r3_scratch
-        cd ~/r3_scratch
-
-	repo init -u git://android.intel.com/manifest -b platform/android/main -m android-main
-        wget $R3_MANIFEST
-        MANIFEST="ls *.xml"
-        sed -i 's/jfumg-gcrmirror.jf.intel.com/ncsgit001.nc.intel.com/g' ./$MANIFEST
-        cp  $MANIFEST ./.repo/manifests
-        repo init -m  $MANIFEST
-        repo sync
-        vendor/intel/support/build_all.sh -c mfld_pr2
 
 
 }
+
 usage(){
 	echo 	" usage is:
 		where x is r2 || r3
@@ -260,7 +293,39 @@ usage(){
 			example:
 				mki.sh -fk r3
 				mki.sh -fk r2
+		-s_x) get code and build EVERYTHING!
+			example:
+				mki.sh -s_r3
+				mki.sh -s_r2
+
+			requires these set on mki.sh, check inside for example!
+			R2_RELEASE
+			R2_URL
+			R2_MANIFEST
+					
+		--full_andrid_build x) runs make at top dir
+			example:
+				mki.sh --full_andrid_build r3
+
+		--get_release x) get DB for flash.
+			example:	
+				mki.sh --get_release r3
 		"
+}
+
+make_shit(){
+	cd $R3_DIR
+	source build/envsetup.sh
+	lunch mfld_pr2-eng
+	cd ~/r3/hardware/intel/PRIVATE/platform_test/RtcPingDownloadTester/src/com/intel	
+#	cp ~/BOX/IN/*.java ./
+	rm -rf  $R3_DIR/out/target/product/mfld_pr2/system/app/RtcPingDownloadTester.apk
+	mm
+	cd $R3_DIR
+	adb uninstall com.intel
+	adb install  $R3_DIR/out/target/product/mfld_pr2/system/app/RtcPingDownloadTester.apk
+	
+
 }
 
 if [ -z "$1" ]; then
@@ -271,13 +336,30 @@ fi
 while [ ! -z "$1" ]; do
   case $1 in
 	#build
-	-k)build_kernel $2;
-		shift;
-		;;
-	-fk)flash_kernel $2;
-		shift;
-		;;	
-	*)usage;;
+	-k)
+		build_kernel $2;
+		break;;
+	-fk)
+		flash_kernel $2;
+		break;;
+	--get_release)
+		get $2;
+		break;;
+	--full_andrid_build)
+		full_andrid_build $2;
+		break;;
+	-s_r3)
+		scratch_r3
+		break;;
+	-s_r2)
+		scratch_r2
+		break;;
+	-v)
+		make_shit
+		break;;	
+	*)
+		usage
+		break;;
 
   esac
   shift
